Compiles but with some issues
Removed encryption/decryption of pokemon substructures for debugging
Debugging when loading next map section

+++ src/daycare.c
@@ -1089,7 +1089,7 @@ void CreateEgg(struct Pokemon *mon, u16 species, bool8 setHotSpringsLocation)
     u8 metLevel;
     u16 ball;
     u8 language;
-    u8 metLocation;
+    u16 metLocation;
     u8 isEgg;
 
     CreateMon(mon, species, EGG_HATCH_LEVEL, 32, FALSE, 0, OT_ID_PLAYER_ID, 0);
